+++ starkinfra/issuingpurchase/__issuingpurchase.py
@@ -1,6 +1,8 @@
+from json import dumps
 from starkcore.utils.resource import Resource
 from starkcore.utils.checks import check_datetime, check_date
 from ..utils import rest
+from ..utils.parse import parse_and_verify
 
 
 class IssuingPurchase(Resource):
@@ -11,6 +13,7 @@ class IssuingPurchase(Resource):
     - holder_name [string]: card holder name. ex: "Tony Stark"
     - card_id [string]: unique id returned when IssuingCard is created. ex: "5656565656565656"
     - card_ending [string]: last 4 digits of the card number. ex: "1234"
+    - purpose [string]: purchase purpose. ex: "purchase"
     - amount [integer]: IssuingPurchase value in cents. Minimum = 0. ex: 1234 (= R$ 12.34)
     - tax [integer]: IOF amount taxed for international purchases. ex: 1234 (= R$ 12.34)
     - issuer_amount [integer]: issuer amount. ex: 1234 (= R$ 12.34)
@@ -28,22 +31,28 @@ class IssuingPurchase(Resource):
     - wallet_id [string]: virtual wallet ID. ex: "5656565656565656"
     - method_code [string]: method code. ex: "chip", "token", "server", "manual", "magstripe" or "contactless"
     - score [float]: internal score calculated for the authenticity of the purchase. None in case of insufficient data. ex: 7.6
-    - issuing_transaction_ids [string]: ledger transaction ids linked to this Purchase
     - end_to_end_id [string]: Unique id used to identify the transaction through all of its life cycle, even before the purchase is denied or accepted and gets its usual id. Example: endToEndId="679cd385-642b-49d0-96b7-89491e1249a5"
-    - status [string]: current IssuingCard status. ex: "approved", "canceled", "denied", "confirmed", "voided"
     - tags [string]: list of strings for tagging returned by the sub-issuer during the authorization. ex: ["travel", "food"]
+    ## Attributes (IssuingPurchase only):
+    - issuing_transaction_ids [string]: ledger transaction ids linked to this Purchase
+    - status [string]: current IssuingCard status. ex: "approved", "canceled", "denied", "confirmed", "voided"
     - updated [datetime.datetime]: latest update datetime for the IssuingPurchase. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
     - created [datetime.datetime]: creation datetime for the IssuingPurchase. ex: datetime.datetime(2020, 3, 10, 10, 30, 0, 0)
+    ## Attributes (authorization request only):
+    - is_partial_allowed [bool]: true if the the merchant allows partial purchases. ex: False
+    - card_tags [list of strings]: tags of the IssuingCard responsible for this purchase. ex: ["travel", "food"]
+    - holder_tags [list of strings]: tags of the IssuingHolder responsible for this purchase. ex: ["technology", "john snow"]
     """
 
-    def __init__(self, id, holder_name, card_id, card_ending, amount, tax, issuer_amount, issuer_currency_code,
+    def __init__(self, id, holder_name, card_id, card_ending, purpose, amount, tax, issuer_amount, issuer_currency_code,
                  issuer_currency_symbol, merchant_amount, merchant_currency_code, merchant_currency_symbol,
                  merchant_category_code, merchant_country_code, acquirer_id, merchant_id, merchant_name, merchant_fee,
-                 wallet_id, method_code, score, issuing_transaction_ids, end_to_end_id, status, tags, created, updated):
+                 wallet_id, method_code, score, end_to_end_id, tags, issuing_transaction_ids, status, updated, created):
         Resource.__init__(self, id=id)
 
         self.holder_name = holder_name
         self.card_id = card_id
+        self.purpose = purpose
         self.card_ending = card_ending
         self.amount = amount
         self.tax = tax
@@ -62,10 +71,10 @@ class IssuingPurchase(Resource):
         self.wallet_id = wallet_id
         self.method_code = method_code
         self.score = score
+        self.tags = tags
         self.issuing_transaction_ids = issuing_transaction_ids
         self.end_to_end_id = end_to_end_id
         self.status = status
-        self.tags = tags
         self.updated = check_datetime(updated)
         self.created = check_datetime(created)
 
@@ -149,3 +158,47 @@ def page(end_to_end_ids=None, holder_ids=None, card_ids=None, status=None, after
         status=status,
         user=user,
     )
+
+
+def parse(content, signature, user=None):
+    """# Create single verified IssuingPurchase authorization request from a content string
+    Use this method to parse and verify the authenticity of the authorization request received at the informed endpoint.
+    Authorization requests are posted to your registered endpoint whenever IssuingPurchases are received.
+    They present IssuingPurchase data that must be analyzed and answered with approval or declination.
+    If the provided digital signature does not check out with the StarkInfra public key, a stark.exception.InvalidSignatureException will be raised.
+    If the authorization request is not answered within 2 seconds or is not answered with an HTTP status code 200 the IssuingPurchase will go through the pre-configured stand-in validation.
+    ## Parameters (required):
+    - content [string]: response content from request received at user endpoint (not parsed)
+    - signature [string]: base-64 digital signature received at response header "Digital-Signature"
+    ## Parameters (optional):
+    - user [Organization/Project object, default None]: Organization or Project object. Not necessary if starkinfra.user was set before function call
+    ## Return:
+    - Parsed IssuingPurchase object
+    """
+    return parse_and_verify(
+        content=content,
+        signature=signature,
+        user=user,
+        resource=_resource,
+        key="",
+    )
+
+
+def response(status, amount=None, reason=None, tags=None):
+    """# Helps you respond IssuingPurchase requests
+    ## Parameters (required):
+    - status [string]: sub-issuer response to the authorization. ex: "approved" or "denied"
+    ## Parameters (conditionally required):
+    - reason [string]: denial reason. Options: "other", "blocked", "lostCard", "stolenCard", "invalidPin", "invalidCard", "cardExpired", "issuerError", "concurrency", "standInDenial", "subIssuerError", "invalidPurpose", "invalidZipCode", "invalidWalletId", "inconsistentCard", "settlementFailed", "cardRuleMismatch", "invalidExpiration", "prepaidInstallment", "holderRuleMismatch", "insufficientBalance", "tooManyTransactions", "invalidSecurityCode", "invalidPaymentMethod", "confirmationDeadline", "withdrawalAmountLimit", "insufficientCardLimit", "insufficientHolderLimit"
+    ## Parameters (optional):
+    - amount [integer, default 0]: amount in cents that was authorized. ex: 1234 (= R$ 12.34)
+    - tags [list of strings, default []]: tags to filter retrieved object. ex: ["tony", "stark"]
+    ## Return:
+    - Dumped JSON string that must be returned to us on the IssuingPurchase request
+    """
+    return dumps({"authorization": {
+        "status": status,
+        "amount": amount or 0,
+        "reason": reason or "",
+        "tags": tags or [],
+    }})