+++ CHANGELOG.md
@@ -15,8 +15,11 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - CreditNotePreview sub-resource
+- response method to PixRequest, PixReversal and IssuingPurchase resources
+- parse method to IssuingPurchase resource
 ### Removed
 - updated attribute from IssuingBin resource
+- IssuingAuthorization resource
 ### Changed
 - fine and interest attributes to return only on CreditNote.Invoice sub-resource